requirements.txt
flask
google-cloud-aiplatform
protobuf